+++ views/cats/index.php
@@ -11,7 +11,7 @@ use wdmg\widgets\SelectInput;
 /* @var $model wdmg\blog\models\Categories */
 
 $this->title = Yii::t('app/modules/blog', 'All categories');
-$this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/blog', 'Blog'), 'url' => ['posts/index']];
+$this->params['breadcrumbs'][] = ['label' => Yii::t('app/modules/blog', 'Blog'), 'url' => ['cats/index']];
 $this->params['breadcrumbs'][] = $this->title;
 
 if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsset')) {
@@ -44,9 +44,24 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                     $output .= (($model->id === 1) ? " <span class=\"text-muted\">(" . Yii::t('app/modules/blog', 'default') . ")</span>" : "");
                     if (($categoryURL = $model->getCategoryUrl(true, true)) && $model->id) {
                         $output .= '<br/>' . Html::a($model->getUrl(true), $categoryURL, [
-                                'target' => '_blank',
-                                'data-pjax' => 0
+                            'target' => '_blank',
+                            'data-pjax' => 0
+                        ]);
+                    }
+
+                    if (isset(Yii::$app->redirects) && $model->url && ($model->status == $model::STATUS_PUBLISHED)) {
+                        if ($url = Yii::$app->redirects->check($model->url, false)) {
+                            $output .= '&nbsp' . Html::tag('span', '', [
+                                'class' => "text-danger fa fa-exclamation-circle",
+                                'data' => [
+                                    'toggle' => "tooltip",
+                                    'placement' => "top"
+                                ],
+                                'title' => Yii::t('app/modules/redirects', 'For this URL is active redirect to {url}', [
+                                    'url' => $url
+                                ])
                             ]);
+                        }
                     }
                     return $output;
                 }
@@ -109,7 +124,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                 'format' => 'html',
                 'value' => function($data) {
                     if ($posts = $data->posts) {
-                        return Html::a(count($posts), ['posts/index', 'cat_id' => $data->id]);
+                        return Html::a(count($posts), ['cats/index', 'cat_id' => $data->id]);
                     } else {
                         return 0;
                     }
@@ -151,7 +166,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                                 if ($data->locale === $locale['locale']) // It`s source version
                                     $output[] = Html::a($flag,
                                         [
-                                            'posts/update', 'id' => $data->id
+                                            'cats/update', 'id' => $data->id
                                         ], [
                                             'title' => Yii::t('app/modules/blog','Edit source version: {language}', [
                                                 'language' => $locale['name']
@@ -161,7 +176,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                                 else  // Other localization versions
                                     $output[] = Html::a($flag,
                                         [
-                                            'posts/update', 'id' => $data->id,
+                                            'cats/update', 'id' => $data->id,
                                             'locale' => $locale['locale']
                                         ], [
                                             'title' => Yii::t('app/modules/blog','Edit language version: {language}', [
@@ -186,7 +201,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                                 if ($data->locale === $locale) // It`s source version
                                     $output[] = Html::a($language,
                                         [
-                                            'news/update', 'id' => $data->id
+                                            'cats/update', 'id' => $data->id
                                         ], [
                                             'title' => Yii::t('app/modules/blog','Edit source version: {language}', [
                                                 'language' => $language
@@ -196,7 +211,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                                 else  // Other localization versions
                                     $output[] = Html::a($language,
                                         [
-                                            'news/update', 'id' => $data->id,
+                                            'cats/update', 'id' => $data->id,
                                             'locale' => $locale
                                         ], [
                                             'title' => Yii::t('app/modules/blog','Edit language version: {language}', [